refactor(digital-screen): establish PR0 frontend boundaries
