+++ yak-ops-ui/src/pages/digital-screen/shared/README.md
@@ -1,0 +1,5 @@
+# Digital Screen Shared
+
+Only contracts, constants or helpers that are genuinely shared by list, editor, viewer and runtime
+belong here. Role-specific code should stay in its feature directory instead of becoming a generic
+`utils` bucket.
