+++ yak-ops-ui/src/services/dataset/index.ts
@@ -1,3 +1,2 @@
 export * from './api';
-export type * from './repository';
 export type * from './types';
