+++ yak-ops-ui/src/services/dataset/types.ts
@@ -1,0 +1,9 @@
+export type {
+  Aggregation,
+  DatasetField,
+  DatasetFieldType,
+  DatasetQueryPayload,
+  DatasetQueryResult,
+  PublishedDataset,
+  Scalar,
+} from '@/components/analysis/model';
