+++ yak-ops-ui/src/services/digital-screen/api.ts
@@ -1,167 +1,18 @@
-import { resolveScreenTemplateById } from '@/services/screen-template-service';
-import type {
-  CreateDigitalScreenInput,
-  DigitalScreenBindings,
-  DigitalScreenInstance,
-  UpdateDigitalScreenInput,
-} from './types';
+import { localScreenRepository } from './repository';
+import type { ScreenRepository } from './repository';
+import type { CreateDigitalScreenInput, UpdateDigitalScreenInput } from './types';
 
-const STORAGE_KEY = 'yak-ops:digital-screens:v1';
+/**
+ * Stable application-facing façade. PR 1 can replace this repository binding with
+ * an HTTP implementation without changing list/editor/viewer callers.
+ */
+export const screenRepository: ScreenRepository = localScreenRepository;
 
-const now = () => new Date().toISOString();
-
-const createId = () => {
-  if (typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function') {
-    return crypto.randomUUID();
-  }
-  return `screen-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`;
-};
-
-const normalizeBindings = (value: unknown): DigitalScreenBindings => {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return {};
-  return value as DigitalScreenBindings;
-};
-
-const readScreens = (): DigitalScreenInstance[] => {
-  if (typeof window === 'undefined') return [];
-  try {
-    const raw = window.localStorage.getItem(STORAGE_KEY);
-    if (!raw) return [];
-    const parsed = JSON.parse(raw) as unknown;
-    if (!Array.isArray(parsed)) return [];
-    return parsed.flatMap((item) => {
-      if (!item || typeof item !== 'object') return [];
-      const candidate = item as Partial<DigitalScreenInstance>;
-      if (
-        typeof candidate.id !== 'string'
-        || typeof candidate.name !== 'string'
-        || typeof candidate.templateId !== 'string'
-      ) return [];
-      return [{
-        ...candidate,
-        bindings: normalizeBindings(candidate.bindings),
-      } as DigitalScreenInstance];
-    });
-  } catch {
-    return [];
-  }
-};
-
-const writeScreens = (screens: DigitalScreenInstance[]) => {
-  if (typeof window === 'undefined') return;
-  window.localStorage.setItem(STORAGE_KEY, JSON.stringify(screens));
-};
-
-const sortScreens = (screens: DigitalScreenInstance[]) => [...screens].sort(
-  (left, right) => new Date(right.updatedAt).getTime() - new Date(left.updatedAt).getTime(),
-);
-
-export const listDigitalScreens = async () => sortScreens(readScreens());
-
-export const getDigitalScreen = async (id: string) => {
-  const screen = readScreens().find((item) => item.id === id);
-  if (!screen) throw new Error('数字化大屏不存在或已被删除');
-  return screen;
-};
-
-export const createDigitalScreen = async (input: CreateDigitalScreenInput) => {
-  const name = input.name.trim();
-  if (!name) throw new Error('请输入大屏名称');
-  if (!resolveScreenTemplateById(input.templateId)) throw new Error('所选大屏模板不存在');
-
-  const timestamp = now();
-  const screen: DigitalScreenInstance = {
-    id: createId(),
-    name,
-    description: input.description?.trim() || undefined,
-    templateId: input.templateId,
-    templateVersion: 1,
-    status: 'draft',
-    bindings: input.bindings ?? {},
-    createdAt: timestamp,
-    updatedAt: timestamp,
-  };
-
-  writeScreens([screen, ...readScreens()]);
-  return screen;
-};
-
-export const updateDigitalScreen = async (id: string, input: UpdateDigitalScreenInput) => {
-  let updated: DigitalScreenInstance | undefined;
-  const screens = readScreens().map((screen): DigitalScreenInstance => {
-    if (screen.id !== id) return screen;
-    const name = input.name === undefined ? screen.name : input.name.trim();
-    if (!name) throw new Error('请输入大屏名称');
-    const next: DigitalScreenInstance = {
-      ...screen,
-      name,
-      description: input.description === undefined
-        ? screen.description
-        : input.description.trim() || undefined,
-      bindings: input.bindings === undefined ? screen.bindings : input.bindings,
-      updatedAt: now(),
-    };
-    updated = next;
-    return next;
-  });
-
-  if (!updated) throw new Error('数字化大屏不存在或已被删除');
-  writeScreens(screens);
-  return updated;
-};
-
-export const publishDigitalScreen = async (id: string) => {
-  let published: DigitalScreenInstance | undefined;
-  const timestamp = now();
-  const screens = readScreens().map((screen): DigitalScreenInstance => {
-    if (screen.id !== id) return screen;
-    const next: DigitalScreenInstance = {
-      ...screen,
-      status: 'published',
-      publishedAt: timestamp,
-      updatedAt: timestamp,
-    };
-    published = next;
-    return next;
-  });
-
-  if (!published) throw new Error('数字化大屏不存在或已被删除');
-  writeScreens(screens);
-  return published;
-};
-
-export const unpublishDigitalScreen = async (id: string) => {
-  let draft: DigitalScreenInstance | undefined;
-  const screens = readScreens().map((screen): DigitalScreenInstance => {
-    if (screen.id !== id) return screen;
-    const next: DigitalScreenInstance = {
-      ...screen,
-      status: 'draft',
-      publishedAt: undefined,
-      updatedAt: now(),
-    };
-    draft = next;
-    return next;
-  });
-
-  if (!draft) throw new Error('数字化大屏不存在或已被删除');
-  writeScreens(screens);
-  return draft;
-};
-
-export const duplicateDigitalScreen = async (id: string) => {
-  const source = await getDigitalScreen(id);
-  return createDigitalScreen({
-    name: `${source.name} - 副本`,
-    description: source.description,
-    templateId: source.templateId,
-    bindings: source.bindings,
-  });
-};
-
-export const deleteDigitalScreen = async (id: string) => {
-  const screens = readScreens();
-  const next = screens.filter((screen) => screen.id !== id);
-  if (next.length === screens.length) throw new Error('数字化大屏不存在或已被删除');
-  writeScreens(next);
-};
+export const listDigitalScreens = () => screenRepository.list();
+export const getDigitalScreen = (id: string) => screenRepository.get(id);
+export const createDigitalScreen = (input: CreateDigitalScreenInput) => screenRepository.create(input);
+export const updateDigitalScreen = (id: string, input: UpdateDigitalScreenInput) => screenRepository.update(id, input);
+export const publishDigitalScreen = (id: string) => screenRepository.publish(id);
+export const unpublishDigitalScreen = (id: string) => screenRepository.unpublish(id);
+export const duplicateDigitalScreen = (id: string) => screenRepository.duplicate(id);
+export const deleteDigitalScreen = (id: string) => screenRepository.remove(id);
