+++ yak-ops-ui/src/services/digital-screen/repository/index.ts
@@ -1,0 +1,2 @@
+export { localScreenRepository } from './local-screen-repository';
+export type { ScreenRepository } from './screen-repository';
